[performance] Schedule resolveAddress immediately
resolveAddress requests are done with one at a time, in order to
correlate the results with the looked-up addresses. Doing them
sequentially via UpdateRequest events added a lot of overhead
in the form of coordination between threads.

This commit creates a QContactFetchRequest for each resolveAddress
call, so that the events for handling them can arrive all together
in the backend thread and the events announcing the results
can arrive all together in the UI thread. They will still be processed
sequentially, but this change cuts out the delay between finishing
one request and starting the next.